backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg[binary]==3.2.9
redis==5.0.4
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
